Performance and Syntax Improvements
* Main script and demo script improved with pylint in Visual Studio Code (Insiders)
* No changes if you already used this project - everything works fine - only faster and better

Note: I think that now succeeded in letting the script eat everything (emojis, äöüÖÜÄß) and nothing goes down - but be aware that especially if you use Python 2 some characters on some devices will be lost on the way. Because I couldn't find out why on some it works and on others not I inserted a try catch method - if your device can do it, it will happen else eveything will still work but information will be lost on the way.